+++ src/form-elements/FormElementCamera.tsx
@@ -2,6 +2,7 @@ import * as React from 'react'
 import clsx from 'clsx'
 import loadImage from 'blueimp-load-image'
 import SignatureCanvas from 'react-signature-canvas'
+import watermark from 'watermarkjs'
 
 import useBooleanState from '../hooks/useBooleanState'
 import downloadFile from '../services/download-file'
@@ -29,6 +30,19 @@ declare global {
 declare global {
   interface Window {
     Camera: any
+  }
+}
+
+function addTextBackground() {
+  return (canvas: HTMLCanvasElement) => {
+    const context = canvas.getContext('2d')
+    if (context) {
+      context.save()
+      context.fillStyle = "rgba(20, 20, 20, 0.6)"
+      context.fillRect(canvas.width - 200, canvas.height -35, canvas.width, canvas.height - 10)
+      context.restore()
+    }
+    return canvas
   }
 }
 function FormElementCamera({
@@ -66,16 +80,24 @@ function FormElementCamera({
         // and someone else has already solved with this nice library.
         // https://nsulistiyawan.github.io/2016/07/11/Fix-image-orientation-with-Javascript.html
         const file = changeEvent.target.files[0]
-        loadImage.parseMetaData(file, function (data) {
+        loadImage.parseMetaData(file, async function (data) {
           const options = {
             // should be set to canvas : true to activate auto fix orientation
             canvas: true,
             // @ts-expect-error if exif data available, update orientation
             orientation: data.exif ? data.exif.get('Orientation') : 0,
           }
+          let blob = file
+          if (element.includeTimestampWatermark) {
+            const now = new Date()
+            blob = await watermark([file])
+            .blob(addTextBackground())
+            .render()
+            .blob(watermark.text.lowerRight(now.toLocaleString(), '20px Arial', '#FFF', 0.85))
+          }
           console.log('Loading image onto canvas to correct orientation')
           loadImage(
-            file,
+            blob,
             (canvas) => {
               // @ts-expect-error this it always be a HTMLCanvasElement because we passed `canvas: true` above
               const base64data = canvas.toDataURL(file.type)